samples/python/voice-live-agent/server/pyproject.toml
[project]
name = "voice-live-agent"
version = "0.1.0"
description = "Real-time voice agent with Azure Voice Live and ACS/Web streaming clients"
dependencies = [
    "quart",
    "websockets",
    "httpx",
    "python-dotenv",
    "quart-cors",
    "sounddevice",
    "numpy",
    "azure-core",
    "azure-communication-identity",
    "aiohttp",
    "azure-eventgrid",
    "azure-identity",
    "azure-communication-callautomation",
    "openai[realtime]"
]
requires-python = ">=3.9"
